https://www.figma.com/proto/5oKKZW3HnoBwKTbulvueYk/ecommerce-app-design?page-id=1%3A2&node-id=51-69&p=f&viewport=-9%2C108%2C0.58&t=F3kMu5AOLxGWAEHc-1&scaling=scale-down&content-scaling=fixed&starting-point-node-id=1%3A3
USER FLOW 
User taps the "Add to Cart" button to include the selected item in their shopping cart. 
A visually appealing success pop-up appaears confirming the successful addition to the cart 
Instead of immediately redirecting to the cart the a smoothly transitions to the next screen. 
The user receives visual cues such as an updated cart icon providing feedback on the current status of the shoin cart 
The user has the option to delete the item if they want.

EXPECTED OUTCOMES 
Seamless Addition: 
Ensure a seamless process when tapping the "Add to Cart" button for including the selected item in the shopping cart. 
Visual Confirmation: 
Display a visually appealing success pop-up to confirm the successful addition to the cart. 
Smooth Transition: 
Facilitate a smooth transition to the next screen instead of an immediate redirection to the cart, maintaining a cohesive and engaging user experience. 
Visual Feedback: 
Provide visual cues, such as an updated cart icon, to offer feedback on the current status of the shopping cart. 
User Control: 
Grant users the option to delete the added item, allowing them control over the contents of their shopping cart